Christmas commit. :-)
I'm having hard times with these theories. It is
really big and my laptop is really hot and not
going much further in the proofs. I'm getting Exception Size Error
I was suggested by Makarius to use the 2016-1 version with 64bit
enabled. At least I god rid of the exception.

Now he suggested me to rethink about the implementation, because
the translation too direct might be unfeasible. So I'm trying to
make a subset of the entire AST/omega functions trying to identify
how far can we go with the current version of our translation.

I'm posting a brief explanation of our work to the Isabelle mailing
list hoping for some feedback from the experts.

Ho! Ho! Ho!