Happy new year!
I had the Z AST shortened a bit - some bits are left out for testing,
and it seems that the size of the AST is way too big for Isabelle (at
least with my computer), and might be one of the reasons I got erros
of lack of resources.
The other reason, I presume, is the use of replicated operators (int
choice, ext choice, sequential composition). If I leave them commented,
Isabelle is able to load it in a reasonable time (seconds). However,
my laptop gets too hot if I leave the replicated operators uncommented,
and at some point I give up and close Isabelle.

Now, I'll write a bit of what I did so far, how I imported my code into
Isabelle (with haskabelle) and the problems I had so far.

Now, off to the beach. Yay!