New Version!!!
I'm pushing a new version of the translator:

Change log:

1 - The previous version was using a list of triple as a parameter for each Omega function,
which would carry the name of each state component, followed by its process name and type.
However, I took the past 10 days to remove that list from the functions. What I did, instead
is to preprocess the spec and rename every state variable with the prefix "st_var_". That will
make the proofs easier as the original mapping functions does not have such feature.

2 - I worked with the auxiliary functions (DefSets.lhs and OmegaDefs.hs), mainly a review
of what was being done by each pattern, and removing any redundancy.

3 - I'm including any function for lists as I can't import libraries with haskabelle - those
functions are included directly in the auxiliary files (OmegaDefs.hs and DefSets.lhs)

4 - The AST (for haskabelle folder) is a bit shorter - see previous commits.

5 - The replicated operators are not very welcome by Isabelle. Huge CPU usage, and I suspect
it is related to functions such rep_CSPRepSeq and rep_CSPRepExtChoice. I even restricted
those for lists of only variables or numbers, so it won't need to evaluate every item of ZExpr.

6 - If I leave the replicated operators commented, I get around 12881 goals for Isabelle to discharge
whilst loading the definition of omega_CAction and omega_prime_CAction. However, if I switch the
comments and leave only the replicated operators and CSPSkip, it almost double it for up to 21k goals.
And I'm pretty sure that the recursion of rep_* functions is killing the prover brain (and mine as well).

Extra: I'm planning to use Isabelle with the SCSS clusters when I'm back to Dublin - can't use it from here.