New version 0.5
Now that we can load the files without the extension .tex, we produce
two other files : .hc and .csp on the same folder of the .tex file:
    .hc is the file containing the haskell AST representation of the spec
    .csp is the translated specification in CSP

Now we also load multiple processes within the same file without name
clashing.

We also are able to have tailored Memory and Memorise actions within
a CSP process with respect only to the types used within the scope
of that process.