New version, new stuff:
Now every variable in the AST has a third field, its type. The translator
will be in charge of filling all those fields with the respective type
of the variable.

Moreover, now we have a correct implementation of the multiple bindings
since the Circus specification. What it was doing before was to
translate into CSP with the definitions being made directly into the CSP text.
Now it comes from the AST and the mapping functions

There are tons of new local functions, still need to comment them, and organize them as well.