Roadwork ahead:
I'm working on the bug from Andrew's comments where the translation should use the parameters of a Circus action to replace the used parameter variables to that values.
FOr that I included a "original" command which shows the original spéc when loaded (it was removed by Andrew in one of his commits, but I really need to have it).

Now what I have to do is to rename the circus action within the context of a CActionCommand CValDecl. Then expand it to the main action.

I'm a bit jet lagged and Sofia is at home with fever, I'll go home and see what I can do with her. I'll come back later on...