Z AxDef to Functions in CSPm
Following Andrew's suggestion, I'm translating functions to CSPM. They are generated as Axiomatic definitions in Z,
which so far were not supported by the parser, and there was not translation rules for it.
Now they are translated as functions in CSPm.

Works nice with PK230Demo.tex

Need to make more tests, but so far I'm happy with it.